Add database ID related utils
