+++ src/redux/sagas/motions/createEditDomainMotion.ts
@@ -15,7 +15,7 @@ import {
   CreateDomainMetadataMutationVariables,
   DomainColor,
 } from '~gql';
-import { getMetadataDatabaseId } from '~utils/domains';
+import { getPendingMetadataDatabaseId } from '~utils/databaseId';
 
 import { ActionTypes } from '../../actionTypes';
 import { AllActions, Action } from '../../types/actions';
@@ -205,7 +205,7 @@ function* createEditDomainMotion({
         mutation: CreateDomainMetadataDocument,
         variables: {
           input: {
-            id: getMetadataDatabaseId(colonyAddress, txHash),
+            id: getPendingMetadataDatabaseId(colonyAddress, txHash),
             name: domainName,
             color: domainColor || DomainColor.LightPink,
             description: domainPurpose || '',
@@ -220,7 +220,7 @@ function* createEditDomainMotion({
         mutation: CreateDomainMetadataDocument,
         variables: {
           input: {
-            id: getMetadataDatabaseId(colonyAddress, txHash),
+            id: getPendingMetadataDatabaseId(colonyAddress, txHash),
             name: domainName,
             color: domainColor || domain.metadata.color,
             description: domainPurpose || domain.metadata.description,
